+++ ripozo/__init__.py
@@ -7,7 +7,7 @@ any protocol.
 from ripozo.decorators import apimethod, translate
 from ripozo.resources import fields, restmixins
 from ripozo.resources.relationships.list_relationship import ListRelationship
-from ripozo.resources.relationships.relationship import Relationship
+from ripozo.resources.relationships.relationship import Relationship, FilteredRelationship
 from ripozo.resources.resource_base import ResourceBase
 from ripozo.resources.request import RequestContainer
 from ripozo.utilities import picky_processor